Break up LandingPage components into sub-components
